# backend/api-first/openapi.yaml
openapi: 3.1.0
info:
  title: KOSMO API
  version: "v1"
  description: >-
    API de KOSMO - plataforma low-code de spec-driven development (SDD) asistida por IA.
    Ciclo completo: Descubrimiento -> Características -> Requisitos -> Modelo -> Código.

    Convenciones:
    - Autenticación PKCE (RFC 7636) con JWT RS256 (bearerAuth).
    - Errores RFC 7807 (application/problem+json) con type URN, trace_id y violations, o OAuthErrorResponse (RFC 6749) / HttpErrorResponse.
    - Fechas ISO-8601 UTC con sufijo Z.
    - Las fases siguen el orden Descubrimiento -> Características -> Requisitos -> Modelo -> Código.
      La consistencia se evalúa exclusivamente hacia la derecha (downstream).
  contact:
    name: Equipo KOSMO
    email: dev@kosmo.app
    url: https://github.com/CesarPantoja1/KOSMO
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local — desarrollo en máquina del programador
  - url: https://api-dev.kosmo.app
    description: Desarrollo — entorno de integración continua
  - url: https://api.kosmo.app
    description: Producción — tráfico real de usuarios
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/projects/{project_id}/discovery/chat:
    post:
      tags: [discovery]
      summary: Enviar mensaje al chat de Descubrimiento (aplicacion instantanea)
      description: >-
        Procesa el mensaje y aplica el cambio inmediatamente sobre el documento cuando corresponde,
        verificando consistencia downstream. Si corresponde a otra fase, devuelve redireccion.
      operationId: discovery_chat_send
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Mensaje procesado, modificacion aplicada o redireccion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Error de validacion o tamano de mensaje
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflicto de transicion de fase
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "502":
          description: Error al invocar el LLM
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    get:
      tags: [discovery]
      summary: Obtener historial de chat de Descubrimiento
      description: Devuelve los mensajes del chat interactivo para la fase de Descubrimiento.
      operationId: discovery_chat_history
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CursorBefore"
        - $ref: "#/components/parameters/SessionIdQuery"
      responses:
        "200":
          description: Historial obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/projects/{project_id}/discovery/chat/stream:
    post:
      tags: [discovery]
      summary: Enviar mensaje al chat de Descubrimiento con streaming SSE
      description: Genera respuesta en tiempo real (SSE). Emite eventos start, chunk y message.
      operationId: discovery_chat_stream
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Stream de eventos SSE
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatStreamEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"

  # ══════════════════════════════ FEATURES ══════════════════════════════
  /api/v1/projects/{project_id}/features:
    post:
      tags: [features]
      summary: Generar caracteristicas del producto con IA
      description: Genera caracteristicas de alto nivel evaluando el documento de descubrimiento del proyecto.
      operationId: features_generate
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Caracteristicas generadas exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [features]
      summary: Listar caracteristicas del proyecto
      description: Devuelve todas las caracteristicas asociadas a un proyecto.
      operationId: features_list
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Lista de caracteristicas del proyecto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/projects/{project_id}/features/{feature_id}:
    delete:
      tags: [features]
      summary: Eliminar caracteristica
      description: Elimina una caracteristica y todos sus artefactos asociados (requisitos, diagramas).
      operationId: features_delete
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FeatureId"
      responses:
        "200":
          description: Caracteristica eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
                  feature_id:
                    type: string
                required: [status, feature_id]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/features/{feature_id}/chat:
    post:
      tags: [features]
      summary: Enviar mensaje al chat de Caracteristicas (aplicacion instantanea)
      description: >-
        Procesa el mensaje y aplica el cambio inmediatamente sobre la caracteristica (titulo, descripcion, origen),
        verificando consistencia downstream en Requisitos y Modelo. Si pertenece a otra fase, devuelve redireccion.
      operationId: feature_chat_send
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdRateLimitQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Mensaje procesado, modificacion aplicada o redireccion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflicto de fase
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/features/{feature_id}/chat/history:
    get:
      tags: [features]
      summary: Obtener historial del chat de la caracteristica
      description: Devuelve todos los mensajes del chat interactivo para una caracteristica especifica.
      operationId: feature_chat_history
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/CursorBefore"
        - $ref: "#/components/parameters/SessionIdQuery"
      responses:
        "200":
          description: Historial obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/features/{feature_id}/chat/stream:
    post:
      tags: [features]
      summary: Enviar mensaje al chat de Caracteristicas con streaming SSE
      description: Genera respuesta en tiempo real (SSE). Emite eventos start, chunk y message.
      operationId: feature_chat_stream
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdRateLimitQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Stream de eventos SSE
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatStreamEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  # ══════════════════════════════ REQUIREMENTS ══════════════════════════════
...
  /api/v1/features/{feature_id}/requirements/chat:
    post:
      tags: [requirements]
      summary: Enviar mensaje al chat de Requisitos (aplicacion instantanea)
      description: >-
        Procesa el mensaje y aplica el cambio inmediatamente sobre el documento de requisitos,
        verificando la consistencia downstream en el Modelo.
      operationId: requirement_chat_send
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdRateLimitQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Mensaje procesado, modificacion aplicada o redireccion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflicto de fase
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/features/{feature_id}/requirements/chat/history:
    get:
      tags: [requirements]
      summary: Obtener historial del chat de un requisito
      description: Devuelve todos los mensajes del chat interactivo para un requisito especifico.
      operationId: requirement_chat_history
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/CursorBefore"
        - $ref: "#/components/parameters/SessionIdQuery"
      responses:
        "200":
          description: Historial obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/features/{feature_id}/requirements/chat/stream:
    post:
      tags: [requirements]
      summary: Enviar mensaje al chat de Requisitos con streaming SSE
      description: Genera respuesta en tiempo real (SSE). Emite eventos start, chunk y message.
      operationId: requirement_chat_stream
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdRateLimitQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Stream de eventos SSE
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatStreamEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  # ══════════════════════════════ MODELO ══════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 emitido por /api/v1/auth/token o /api/v1/auth/refresh
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
      description: Identificador del proyecto (prj_ + ULID)
    FeatureId:
      name: feature_id
      in: path
      required: true
      schema:
        type: string
      description: Identificador o slug de la característica
...
    CursorBefore:
      name: before
      in: query
      required: false
      schema:
        type: string
      description: Cursor ISO-8601 para paginación hacia atrás
    SessionIdQuery:
      name: session_id
      in: query
      required: false
      schema:
        type: string
      description: Hilo de conversación (cht_ + ULID). Si se omite, contexto de la fase.
...
    ProjectIdRateLimitQuery:
      name: project_id
      in: query
      required: false
      schema:
        type: string
        default: ""
      description: Identificador opcional del proyecto para aplicar rate limiting cuando se invoca fuera de una ruta anidada de proyecto.
  responses:
    Unauthorized:
      description: Token de acceso ausente, invalido, expirado o revocado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HttpErrorResponse"
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    BadRequest:
      description: Peticion invalida o error de validacion
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    BadGateway:
      description: Error al invocar el servicio de IA o LLM
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
    ProblemDetail:
      type: object
      description: Error RFC 7807 (application/problem+json)
      properties:
        type:
          type: string
          description: URN del error (ej. urn:kosmo:projects:not-found)
          example: urn:kosmo:projects:not-found
        title:
          type: string
          description: Titulo corto del error
        status:
          type: integer
          description: Codigo de estado HTTP
        detail:
          type: string
          description: Descripcion detallada del error
        instance:
          type: string
          description: Ruta o recurso que origino el error
        trace_id:
          type: string
          description: Identificador de correlacion distribuida (X-Trace-Id)
        violations:
          type: array
          items:
            type: object
            properties:
              loc:
                type: string
                description: Ruta al campo invalido
              msg:
                type: string
                description: Mensaje de error de validacion
              input:
                description: Valor recibido que causo la violacion
      required: [type, title, status, detail]
    HttpErrorResponse:
      type: object
      description: Respuesta de error generica para errores HTTP (4xx/5xx)
      properties:
        detail:
          type: string
          description: Mensaje descriptivo del error
          example: Recurso no encontrado.
      required: [detail]

    # ── Auth ──
...
    FeatureResponse:
      type: object
      properties:
        id:
          type: string
          description: Identificador unico de la caracteristica
          example: feat_01HXYAZABCDEFGHIJKLMNOP
        project_id:
          type: string
          example: prj_01HXYAZABCDEFGHIJKLMNOP
        number:
          type: integer
          example: 1
        title:
          type: string
          example: Gestion de catalogo de productos
        slug:
          type: string
          example: gestion-de-catalogo-de-productos
        description:
          type: string
          example: Permite a los usuarios administrar el catalogo de productos.
        origin:
          type: string
          example: Se deriva de la seccion Capacidades del descubrimiento.
        display_id:
          type: string
          example: C01
        warnings:
          type: array
          items:
            type: string
          description: Advertencias sobre solapamientos funcionales o inconsistencias menores
          default: []
      required: [id, project_id, number, title, slug, description, origin, display_id]
...
    SendChatRequest:
      type: object
      additionalProperties: false
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
          description: Mensaje textual del usuario
        session_id:
          type: string
          nullable: true
          description: Hilo de conversacion (cht_ + ULID). Si se omite, contexto de la fase.
      required: [content]
    ChangeSuggestion:
      type: object
      description: Sugerencia de cambio generada por la IA
      properties:
        id:
          type: string
          description: ID de la sugerencia (chg_ + ULID)
        section:
          type: string
          description: Seccion del documento afectada
        description:
          type: string
          description: Descripcion corta del cambio propuesto
        diff_before:
          type: string
          description: Contenido actual de la seccion
        diff_after:
          type: string
          description: Contenido sugerido por la IA
        rationale:
          type: string
          nullable: true
          description: Justificacion del cambio propuesto
        applied:
          type: boolean
          default: true
          description: True si el servidor aplico el cambio al instante
        not_applied_reason:
          type: string
          nullable: true
          description: Razon por la que no se pudo aplicar
      required: [id, section, description, diff_before, diff_after, applied]
    ModificationResultView:
      type: object
      description: Resultado de la modificacion aplicada por el chat
      properties:
        applied:
          type: boolean
          description: True si la modificacion fue aplicada al documento
        modified_section:
          type: string
          nullable: true
          description: Nombre de la seccion o atributo modificado
        change_description:
          type: string
          nullable: true
          description: Descripcion breve del cambio aplicado
        modified_document:
          type: string
          nullable: true
          description: Markdown completo actualizado
        before:
          type: string
          nullable: true
          description: Contenido previo modificado
        after:
          type: string
          nullable: true
          description: Contenido posterior modificado
        undo_version_id:
          type: string
          nullable: true
          description: Version previa al cambio (reservado para undo)
        clarification_message:
          type: string
          nullable: true
          description: Mensaje solicitando detalle cuando la instruccion es ambigua
      required: [applied]
    RedirectInfoView:
      type: object
      description: Redireccion cuando el cambio pertenece a otra fase upstream
      properties:
        target_phase:
          type: string
          description: Fase destino del cambio
        redirect_message:
          type: string
          description: Explicacion de por que el cambio pertenece a esa fase
      required: [target_phase, redirect_message]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          description: ID del mensaje
        role:
          type: string
          description: Rol del emisor (user, assistant, system)
        content:
          type: string
          description: Contenido del mensaje
        created_at:
          type: string
          format: date-time
        change_suggestions:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ChangeSuggestion"
        modification:
          $ref: "#/components/schemas/ModificationResultView"
      required: [id, role, content, created_at]
    ChatResponse:
      type: object
      description: Respuesta del chat con modificacion aplicada, consistencia o redireccion
      properties:
        message:
          $ref: "#/components/schemas/ChatMessage"
        modification:
          $ref: "#/components/schemas/ModificationResultView"
        redirect:
          $ref: "#/components/schemas/RedirectInfoView"
        consistency:
          type: array
          nullable: true
          items:
            type: object
          description: Impactos de consistencia en las fases downstream
      required: [message]
    ChatHistoryResponse:
      type: object
      properties:
        phase:
          type: string
        context:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        has_more:
          type: boolean
          default: false
        next_cursor:
          type: string
          nullable: true
      required: [phase, context, messages]
    ChatStreamEvent:
      oneOf:
        - type: object
          description: Inicio del stream
          properties:
            type:
              type: string
              const: start
          required: [type]
        - type: object
          description: Fragmento de texto en tiempo real
          properties:
            type:
              type: string
              const: chunk
            content:
              type: string
          required: [type, content]
        - type: object
          description: Mensaje final completo con cards de cambio
          properties:
            type:
              type: string
              const: message
            id:
              type: string
            role:
              type: string
            content:
              type: string
            suggestions:
              type: array
              items:
                $ref: "#/components/schemas/ChangeSuggestion"
            modification:
              $ref: "#/components/schemas/ModificationResultView"
            timestamp:
              type: string
              format: date-time
          required: [type, id, role, content]

    # ── Chat Sessions ──
...